refactor: equals 함수 리팩토링
