+++ packages/assignment/src/@lib/equalities/deepEquals.ts
@@ -1,17 +1,5 @@
-import { isObject, compareArrays, compareObjects } from '@/utils'
+import { shallowEquals } from './shallowEquals'
 
 export function deepEquals(objA: unknown, objB: unknown): boolean {
-  if (objA === objB) {
-    return true
-  }
-
-  if (!isObject(objA) || !isObject(objB)) {
-    return false
-  }
-
-  if (Array.isArray(objA) && Array.isArray(objB)) {
-    return compareArrays(objA, objB, deepEquals)
-  }
-
-  return compareObjects(objA, objB, deepEquals)
+  return shallowEquals(objA, objB, deepEquals)
 }
