+++ packages/assignment/src/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './config'
-export * from './compare'
